+++ src/Tools/dotnet-monitor/GenerateApiKeyCommandHandler.cs
@@ -3,11 +3,8 @@
 // See the LICENSE file in the project root for more information.
 
 using System;
-using System.Collections.Generic;
 using System.CommandLine;
 using System.CommandLine.IO;
-using System.Security.Cryptography;
-using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 
@@ -29,49 +26,6 @@ namespace Microsoft.Diagnostics.Tools.Monitor
             console.Out.WriteLine(FormattableString.Invariant($"ApiKeyHashType: {newKey.HashAlgorithm}"));
 
             return Task.FromResult(0);
-        }
-    }
-
-    internal sealed class GeneratedApiKey
-    {
-        private const int KeyLengthBytes = 32;
-
-        public readonly string MonitorApiKey;
-        public readonly string HashAlgorithm;
-        public readonly string HashValue;
-
-        private GeneratedApiKey(string monitorApiKey, string hashAlgorithm, string hashValue)
-        {
-            this.MonitorApiKey = monitorApiKey;
-            this.HashAlgorithm = hashAlgorithm;
-            this.HashValue = hashValue;
-        }
-
-        public static GeneratedApiKey Create()
-        {
-            return GeneratedApiKey.Create("SHA256");
-        }
-
-        private static GeneratedApiKey Create(string hashAlgorithm)
-        {
-            using RandomNumberGenerator rng = RandomNumberGenerator.Create();
-            using HashAlgorithm hasher = SHA256.Create();
-
-            byte[] secret = new byte[KeyLengthBytes];
-            rng.GetBytes(secret);
-
-            byte[] hash = hasher.ComputeHash(secret);
-            StringBuilder outHash = new StringBuilder(hash.Length * 2);
-
-            foreach (byte b in hash)
-            {
-                outHash.AppendFormat("{0:X2}", b);
-            }
-
-            string apiKey = Convert.ToBase64String(secret);
-
-            GeneratedApiKey result = new GeneratedApiKey(apiKey, hashAlgorithm, outHash.ToString());
-            return result;
         }
     }
 }
